+++ src/undate/converters/combined.py
@@ -1,18 +1,19 @@
 """
-**Experimental** combined parser. Supports EDTF, Hebrew, and Hijri
-where dates are unambiguous. (Year-only dates are parsed as EDTF in
-Gregorian calendar.)
+**Experimental** combined parser. Supports EDTF, Gregorian, Hebrew, and Hijri
+where dates are unambiguous. Year-only dates are parsed as EDTF in
+Gregorian calendar.
 """
 
 from typing import Union
 
 from lark import Lark
-from lark.exceptions import UnexpectedCharacters
+from lark.exceptions import UnexpectedInput
 from lark.visitors import Transformer, merge_transformers
 
 from undate import Undate, UndateInterval
 from undate.converters import BaseDateConverter, GRAMMAR_FILE_PATH
 from undate.converters.edtf.transformer import EDTFTransformer
+from undate.converters.calendars.gregorian.transformer import GregorianDateTransformer
 from undate.converters.calendars.hebrew.transformer import HebrewDateTransformer
 from undate.converters.calendars.islamic.transformer import IslamicDateTransformer
 
@@ -33,6 +34,7 @@ combined_transformer = merge_transformers(
     edtf=EDTFTransformer(),
     hebrew=HebrewDateTransformer(),
     islamic=IslamicDateTransformer(),
+    gregorian=GregorianDateTransformer(),
 )
 
 
@@ -45,7 +47,7 @@ parser = Lark.open(
 class OmnibusDateConverter(BaseDateConverter):
     """
     Combination parser that aggregates existing parser grammars.
-    Currently supports EDTF, Hebrew, and Hijri  where dates are unambiguous.
+    Currently supports EDTF, Gregorian, Hebrew, and Hijri where dates are unambiguous.
     (Year-only dates are parsed as EDTF in Gregorian calendar.)
 
     Does not support serialization.
@@ -75,7 +77,7 @@ class OmnibusDateConverter(BaseDateConverter):
             parsetree = parser.parse(value)
             # transform returns a list; we want the first item in the list
             return self.transformer.transform(parsetree)[0]
-        except UnexpectedCharacters:
+        except UnexpectedInput:
             raise ValueError(
                 "Parsing failed: '%s' is not in a recognized date format" % value
             )